add json schema for Theme type
