+++ src/contexts/ThemeContext.tsx
@@ -1,4 +1,5 @@
-import { type Theme, applyTheme } from '@/lib/theme';
+import { applyTheme } from '@/lib/theme';
+import { Theme } from '@/lib/theme.type';
 import { getThemeByName, invalidateThemeCache, loadThemes } from '@/lib/themes';
 import React, { createContext, useContext, useEffect, useState } from 'react';
 import { useLocalStorage } from 'usehooks-ts';
